Format all Python files consistently
The files in this repository currently follow no formatting standards,
which causes inconsistencies in things like using tabs or spaces for
indentation. We use [yapf](https://github.com/google/yapf) to format
every file in the repository with the following command:

```bash
yapf -i -r .
```

This normalizes all of the code in the repository to follow a single
standard and also makes it easy for future contributors to ensure that
they are following the formatting standards (they simply have to run
yapf themselves before committing).

Sorry if this commit causes a lot of churn to the code, but it should be
safe given that yapf is very battle-tested and has been used for
countless other Python projects.